[FIX] Moved resume regulation scripts in another script.
